Renamed index.js to processor.js
